requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
PyMySQL==1.0.3
cryptography==41.0.7
Werkzeug==2.3.7
reportlab==4.0.4
openpyxl==3.1.2
pandas==2.1.4
matplotlib==3.7.2